Todos by user

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse } from "next/server";
 import * as Yup from "yup";
@@ -43,12 +44,17 @@ const postSchema = Yup.object({
 });
 
 export async function POST(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+
   try {
     const { complete, description } = await postSchema.validate(
       await request.json()
     );
 
-    const todo = await prisma.todo.create({ data: { complete, description } });
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } });
 
     return NextResponse.json(todo);
   } catch (error) {
@@ -57,10 +63,15 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
   try {
     await prisma.todo.deleteMany({
       where: {
         complete: true,
+        userId: user.id,
       },
     });
     return NextResponse.json(
